Conflicts:
	CommonUtils.podspec
	Podfile.lock

+++ CommonUtils/Classes/Categories/UIImage/UIImage+Resize.m
@@ -30,9 +30,9 @@
 
 // Returns a copy of this image that is squared to the thumbnail size.
 // If transparentBorder is non-zero, a transparent border of the given size will be added around the edges of the thumbnail. (Adding a transparent border of at least one pixel in size has the side-effect of antialiasing the edges of the image when rotating it using Core Animation.)
-- (UIImage *)thumbnailImage:(NSInteger)thumbnailSize
-          transparentBorder:(NSUInteger)borderSize
-               cornerRadius:(NSUInteger)cornerRadius
+- (UIImage *)thumbnailImage:(CGFloat)thumbnailSize
+          transparentBorder:(CGFloat)borderSize
+               cornerRadius:(CGFloat)cornerRadius
        interpolationQuality:(CGInterpolationQuality)quality {
     UIImage *resizedImage = [self resizedImageWithContentMode:UIViewContentModeScaleAspectFill
                                                        bounds:CGSizeMake(thumbnailSize, thumbnailSize)
